add autoincrement to strategy indexes

+++ programs/strategy/src/state/base_strategy.rs
@@ -19,7 +19,7 @@ pub trait StrategyInit {
     fn init(
         &mut self, 
         bump: u8,
-        index: u8,
+        index: u64,
         vault: Pubkey, 
         underlying_mint: &InterfaceAccount<Mint>, 
         underlying_token_acc: Pubkey,
